Perfect collision checking.

+++ src/domain/physicalobjects/Ball.java
@@ -9,7 +9,7 @@ import javax.swing.*;
 
 public class Ball extends PhysicalObject{
 
-    public Ball(Vector location, ImageIcon image, int width, int height ){
+    public Ball(Vector location, ImageIcon image, double width, double height ){
         super(location, image, width, height, new SphereBoundingBox(location.add(new Vector(width/2, width/2)), width/2 ), new BallMovementBehavior(), new BallCollisionBehavior());
     }
