+++ src/domain/physicalobjects/Wall.java
@@ -10,7 +10,7 @@ public class Wall extends PhysicalObject{
 
     private Vector normal;
 
-    public Wall(Vector location, int width, int height, Vector normal) {
+    public Wall(Vector location, double width, double height, Vector normal) {
         super(location, null, width, height, new StationaryMovementBehavior(), new NoCollisionBehavior());
         this.normal = normal;
     }
